mempool: rework lock discipline to mitigate callback deadlocks
The priority mempool has a stricter synchronization requirement than the legacy
mempool. In the socket client under sufficiently-heavy load, exclusive access
can lead to deadlocks when processing a large batch of transaction rechecks.

Under the covers, the deadlock occurs because the socket client must respond to
all requests in order of issue from a fixed-size queue. The combination means
that a response callback can pin a subsequent flush from issuing, during a
phase when the mempool lock is held for update.

An ideal fix to this would be to allow the socket client to process responses
in order of completion, but that would disrupt the legacy mempool, which at the
moment depends on the order to correctly handle recheck cleanup.

As a workaround, process rechecks for the priority mempool in their own
goroutines. The socket client will still exhibit head-of-line blocking for
order, but will no longer attempt to acquire a lock to update the state till
their response from the server is complete.